move liquidity components to pool module
