+++ src/modules/pool/components/AddLiquidity/Confirm.vue
@@ -1,5 +1,10 @@
 <template>
-  <dialog-base :visible.sync="isVisible" :title="t('confirmSupply.title')" v-if="firstToken && secondToken">
+  <dialog-base
+    :visible.sync="isVisible"
+    :title="t('confirmSupply.title')"
+    v-if="firstToken && secondToken"
+    append-to-body
+  >
     <div class="pool-tokens-amount">{{ shareOfPool }}%</div>
     <s-row v-if="firstToken && secondToken" flex align="middle" class="pool-tokens">
       <pair-token-logo :first-token="firstToken" :second-token="secondToken" size="small" />
@@ -75,7 +80,7 @@ import { lazyComponent } from '@/router';
     PairTokenLogo: lazyComponent(Components.PairTokenLogo),
   },
 })
-export default class ConfirmAddLiquidity extends Mixins(
+export default class AddLiquidityConfirm extends Mixins(
   mixins.FormattedAmountMixin,
   mixins.LoadingMixin,
   mixins.DialogMixin,
