+++ src/modules/pool/components/AddLiquidity/TransactionDetails.vue
