+++ src/modules/pool/components/RemoveLiquidity/TransactionDetails.vue
